requirements.txt
Flask
Pillow
moviepy
gunicorn
python-dotenv
tqdm
imageio
numpy
proglog
imageio_ffmpeg
decorator
